requirements.txt
gradio==4.29.0
langchain==1.2.15
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-groq==1.1.2
google-search-results==2.4.2
ddgs==9.14.0
python-dotenv==1.2.1
